functions/ingest/requirements.txt
google-cloud-discoveryengine==0.13.4
google-cloud-storage==2.18.0
functions-framework==3.8.1